// generated/src/app/api-models/group-contract.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class GroupContract {
  seqGroupContract: number;
  seqGroupId: number;
  effectiveDate: Date;
  termDate: Date;
  termReason: string;
  brokerName: string;
  salespersonName: string;
  openEnrollStart: Date;
  openEnrollEnd: Date;
  numberOfEmployees: number;
  waitingPeriod: number;
  studentAgeFrom: number;
  studentAgeTo: number;
  contractType: string;
  userDefined1: string;
  userDefined2: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  seqAgentId: number;
  userDefined3: string;
  userDefined4: string;
  userDefined5: string;
  userDefined6: string;
  userDefined7: string;
  userDefined8: string;
  userDefined9: string;
  userDefined10: string;
  userDate1: Date;
  userDate2: Date;
  userDate3: Date;
  userDate4: Date;
  userDate5: Date;
  userDate6: Date;
  userDate7: Date;
  userDate8: Date;
  userDate9: Date;
  userDate10: Date;
}